:fire: Chore(src): cleanup practice
