--- lib/libtensor/vector/Makefile
@@ -1,28 +1,0 @@
-# **************************************************************************** #
-#                                                                              #
-#                                                         :::      ::::::::    #
-#    Makefile                                           :+:      :+:    :+:    #
-#                                                     +:+ +:+         +:+      #
-#    By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+         #
-#                                                 +#+#+#+#+#+   +#+            #
-#    Created: 2023/01/05 08:31:25 by jgo               #+#    #+#              #
-#    Updated: 2023/06/05 20:19:36 by jgo              ###   ########.fr        #
-#                                                                              #
-# **************************************************************************** #
-
-ifndef TOPDIR
-		TOPDIR = $(abspath ../../../)
-endif
-
-include $(TOPDIR)/make_config/com/ColorRules.mk
-include $(TOPDIR)/make_config/lib/Rules.mk
-
-SRCS :=
-
-NAME :=	libvector.a
-
-OBJS := $(SRCS:.c=.o)
-DEPS := $(SRCS:.c=.d)
--include $(DEPS)
-
-include $(TOPDIR)/make_config/lib/SubLibsRecipes.mk
