+++ src/mandatory/includes/enum.h
@@ -6,7 +6,7 @@
 /*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/25 16:12:08 by jgo               #+#    #+#             */
-/*   Updated: 2023/06/05 19:48:56 by jgo              ###   ########.fr       */
+/*   Updated: 2023/06/07 17:04:45 by jgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,6 @@ enum						e_object_type
 	SP,
 	PL,
 	CY
-//LIGHT_POINT
 };
 
 #endif
