+++ src/mandatory/includes/minirt.h
@@ -6,7 +6,7 @@
 /*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/25 15:47:47 by jgo               #+#    #+#             */
-/*   Updated: 2023/06/06 18:36:22 by jgo              ###   ########.fr       */
+/*   Updated: 2023/06/07 16:57:47 by jgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 # include <fcntl.h>
 # include <stdio.h>
 # include <stdlib.h>
-# include <math.h>
 
+# include "vector.h"
 # include "MLX42.h"
 # include "libft.h"
