+++ src/mandatory/includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/27 11:09:46 by jgo               #+#    #+#             */
-/*   Updated: 2023/06/06 18:51:44 by jgo              ###   ########.fr       */
+/*   Updated: 2023/06/07 17:01:56 by jgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,28 +15,7 @@
 
 # include "defs.h"
 
-t_vec3		vec3(double x, double y, double z);
-t_point3	point3(double x, double y, double z);
-t_point3	color3(double r, double g, double b);
-void		vset(t_vec3 *vec, double x, double y, double z);
-double		vlength2(t_vec3 vec);
-double		vlength(t_vec3 vec);
-t_vec3		vplus(t_vec3 vec, t_vec3 vec2);
-t_vec3		vplus_(t_vec3 vec, double x, double y, double z);
-t_vec3		vminus(t_vec3 vec, t_vec3 vec2);
-t_vec3		vminus_(t_vec3 vec, double x, double y, double z);
-t_vec3		vmult(t_vec3 vec, double t);
-t_vec3		vmult_(t_vec3 vec, t_vec3 vec2);
-t_vec3		vdivide(t_vec3 vec, double t);
-double		vdot(t_vec3 vec, t_vec3 vec2);
-t_vec3		vcross(t_vec3 vec, t_vec3 vec2);
-t_vec3		vunit(t_vec3 vec);
-t_vec3		vmin(t_vec3 vec1, t_vec3 vec2);
-
-void		oadd(t_object **list, t_object *new);
-t_object	*olast(t_object *list);
-
-void		destroy(const t_meta *meta);
-void		hooks(void *param);
+void	destroy(const t_meta *meta);
+void	hooks(void *param);
 
 #endif
