+++ src/mandatory/main.c
@@ -1,77 +1,29 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   main.c                                             :+:      :+:    :+:   */
+/*   main_mlx.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/27 11:07:29 by jgo               #+#    #+#             */
-/*   Updated: 2023/05/28 21:57:47 by jgo              ###   ########.fr       */
+/*   Created: 2023/05/25 15:49:19 by jgo               #+#    #+#             */
+/*   Updated: 2023/06/06 18:50:20 by jgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
-#include "print.h"
-#include "scene.h"
-#include "struct.h"
-#include "trace.h"
+#include "defs.h"
+#include "design_patterns.h"
 #include "utils.h"
 
-t_scene	*scene_init(void)
+int	main(int ac, char **av)
 {
-	t_scene		*scene;
-	t_object	*world;
-	t_object	*lights;
+	t_meta	*meta;
 
-	double ka; // 8.4 에서 설명
-	// malloc 할당 실패 시, 실습에서는 return NULL로 해두었지만, 적절한 에러 처리가 필요하다.
-	if (!(scene = (t_scene *)malloc(sizeof(t_scene))))
-		return (NULL);
-	scene->canvas = canvas(400, 300);
-	scene->camera = camera(&scene->canvas, point3(0, 0, 0));
-	world = object(SP, sphere(point3(-2, 0, -5), 2), color3(0.5, 0, 0));
-	// world 에 구1 추가
-	oadd(&world, object(SP, sphere(point3(0, -1000, 0), 995), color3(1, 1, 1)));
-	// world 에 구2 추가
-	oadd(&world, object(SP, sphere(point3(2, 0, -5), 2), color3(0, 0.5, 0)));
-	// world 에 구3 추가
-	scene->world = world;
-	lights = object(LIGHT_POINT, light_point(point3(0, 20, 0), color3(1, 1, 1),
-				0.5), color3(0, 0, 0)); // 더미 albedo
-	scene->light = lights;
-	ka = 0.1;                                   // 8.4 에서 설명
-	scene->ambient = vmult(color3(1, 1, 1), ka); // 8.4 에서 설명
-	return (scene);
-}
-
-int	main(void)
-{
-	int i;
-	int j;
-	double u;
-	double v;
-	t_color3 pixel_color;
-	t_scene *scene;
-
-	scene = scene_init();
-	printf("P3\n%d %d\n255\n", scene->canvas.width, scene->canvas.height);
-	j = scene->canvas.height - 1;
-	while (j >= 0)
-	{
-		i = 0;
-		while (i < scene->canvas.width)
-		{
-			u = (double)i / (scene->canvas.width - 1);
-			v = (double)j / (scene->canvas.height - 1);
-			//ray from camera origin to pixel
-			scene->ray = ray_primary(&scene->camera, u, v);
-			pixel_color = ray_color(scene);
-			// ray_color함수의 인자도 ray, world를 모두 담고 있는 scene으로 바꿨다.
-			/* * * * 수정 끝 * * * */
-			write_color(pixel_color);
-			++i;
-		}
-		--j;
-	}
-	return (0);
+	(void)ac;
+	(void)av;
+	meta = singleton();
+	mlx_image_to_window(meta->mlx_assets.mlx, meta->mlx_assets.img, 0, 0);
+	mlx_loop_hook(meta->mlx_assets.mlx, hooks, meta);
+	mlx_loop(meta->mlx_assets.mlx);
+	return (EXIT_SUCCESS);
 }
