--- src/mandatory/main_mlx.c
@@ -1,29 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   main_mlx.c                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/25 15:49:19 by jgo               #+#    #+#             */
-/*   Updated: 2023/06/06 18:50:20 by jgo              ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minirt.h"
-#include "defs.h"
-#include "design_patterns.h"
-#include "utils.h"
-
-int	main(int ac, char **av)
-{
-	t_meta	*meta;
-
-	(void)ac;
-	(void)av;
-	meta = singleton();
-	mlx_image_to_window(meta->mlx_assets.mlx, meta->mlx_assets.img, 0, 0);
-	mlx_loop_hook(meta->mlx_assets.mlx, hooks, meta);
-	mlx_loop(meta->mlx_assets.mlx);
-	return (EXIT_SUCCESS);
-}
