+++ src/mandatory/parser/parser.c
@@ -3,12 +3,11 @@
 /*                                                        :::      ::::::::   */
 /*   parser.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sanghwal <sanghwal@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/05 14:59:02 by sanghwal          #+#    #+#             */
-/*   Updated: 2023/06/05 14:59:20 by sanghwal         ###   ########seoul.kr  */
+/*   Updated: 2023/06/07 17:04:38 by jgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "defs.h"
-
