--- src/mandatory/print/print.c
@@ -1,21 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   print.c                                            :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/27 11:07:03 by jgo               #+#    #+#             */
-/*   Updated: 2023/05/27 22:58:07 by jgo              ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "print.h"
-
-// [0,1] 로 되어있는 rgb 값을 각각 [0,255]에 맵핑 해서 출력.
-void    write_color(t_color3 pixel_color)
-{
-    printf("%d %d %d\n", (int)(255.999 * pixel_color.x),
-                        (int)(255.999 * pixel_color.y),
-                        (int)(255.999 * pixel_color.z));
-}
