--- src/mandatory/scene/Makefile
@@ -1,30 +1,0 @@
-# **************************************************************************** #
-#                                                                              #
-#                                                         :::      ::::::::    #
-#    Makefile                                           :+:      :+:    :+:    #
-#                                                     +:+ +:+         +:+      #
-#    By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+         #
-#                                                 +#+#+#+#+#+   +#+            #
-#    Created: 2023/01/05 08:31:25 by jgo               #+#    #+#              #
-#    Updated: 2023/05/27 23:09:54 by jgo              ###   ########.fr        #
-#                                                                              #
-# **************************************************************************** #
-
-ifndef TOPDIR
-		TOPDIR = $(abspath ../../../)
-endif
-
-include $(TOPDIR)/make_config/com/ColorRules.mk
-include $(TOPDIR)/make_config/src/Rules.mk
-
-SRCS :=	canvas.c \
-		scene.c \
-		object_create.c
-
-NAME := scene.a
-
-OBJS := $(SRCS:.c=.o)
-DEPS := $(SRCS:.c=.d)
--include $(DEPS)
-
-include $(TOPDIR)/make_config/src/SubLibsRecipes.mk
