--- src/mandatory/scene/object_create.c
@@ -1,53 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   object_create.c                                    :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/27 19:26:51 by jgo               #+#    #+#             */
-/*   Updated: 2023/05/28 17:49:07 by jgo              ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "struct.h"
-/* * * * 수정 * * * */
-#include "scene.h"
-
-t_object	*object(t_object_type type, void *element, t_color3 albedo)
-{
-	t_object	*new;
-
-	if (!(new = (t_object *)malloc(sizeof(t_object))))
-		return (NULL);
-	new->type = type;
-	new->element = element;
-	new->next = NULL;
-	new->albedo = albedo;
-	return (new);
-}
-
-t_sphere	*sphere(t_point3 center, double radius)
-{
-	t_sphere	*sp;
-
-	if (!(sp = (t_sphere *)malloc(sizeof(t_sphere))))
-		return (NULL);
-	sp->center = center;
-	sp->radius = radius;
-	sp->radius2 = radius * radius;
-	return (sp);
-}
-
-t_light	*light_point(t_point3 light_origin, t_color3 light_color,
-		double bright_ratio)
-{
-	t_light *light;
-
-	if (!(light = (t_light *)malloc(sizeof(t_light))))
-		return (NULL);
-	light->origin = light_origin;
-	light->light_color = light_color;
-	light->bright_ratio = bright_ratio;
-	return (light);
-}
