--- src/mandatory/scene/scene.c
@@ -1,35 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   scene.c                                            :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/27 17:24:38 by jgo               #+#    #+#             */
-/*   Updated: 2023/05/27 22:58:07 by jgo              ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "scene.h"
-
-t_camera	camera(t_canvas *canvas, t_point3 orig)
-{
-	t_camera cam;
-	double focal_len;
-	double viewport_height;
-
-	viewport_height = 2.0;
-	focal_len = 1.0;
-	cam.orig = orig;
-	cam.viewport_h = viewport_height;
-	cam.viewport_w = viewport_height * canvas->aspect_ratio;
-	cam.focal_len = focal_len;
-	cam.horizontal = vec3(cam.viewport_w, 0, 0);
-	cam.vertical = vec3(0, cam.viewport_h, 0);
-	// 왼쪽 아래 코너점 좌표, origin - horizontal / 2 - vertical / 2 - vec3(0,0,focal_length)
-	cam.left_bottom = vminus(vminus(vminus(cam.orig, vdivide(cam.horizontal,
-						2)),
-									vdivide(cam.vertical, 2)),
-								vec3(0, 0, focal_len));
-	return (cam);
-}
