+++ src/mandatory/utils/hooks.c
@@ -6,18 +6,19 @@
 /*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/25 17:08:41 by jgo               #+#    #+#             */
-/*   Updated: 2023/06/06 18:51:55 by jgo              ###   ########.fr       */
+/*   Updated: 2023/06/07 17:04:35 by jgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minirt.h"
 #include "defs.h"
+#include "minirt.h"
 #include "utils.h"
 
 void	hooks(void *param)
 {
-	const t_meta *meta = param;
+	const t_meta	*meta;
 
+	meta = param;
 	if (mlx_is_key_down(meta->mlx_assets.mlx, MLX_KEY_ESCAPE))
 		destroy(meta);
 }
