--- src/mandatory/utils/object_utils.c
@@ -1,38 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   object_utils.c                                     :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/27 23:33:10 by jgo               #+#    #+#             */
-/*   Updated: 2023/05/27 23:33:26 by jgo              ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "defs.h"
-#include "utils.h"
-
-void	oadd(t_object **list, t_object *new)
-{
-	t_object	*cur;
-
-	if (*list == NULL)
-	{
-		*list = new;
-		return ;
-	}
-	cur = *list;
-	while (cur->next)
-		cur = cur->next;
-	cur->next = new;
-}
-
-t_object	*olast(t_object *list)
-{
-	if (list == NULL)
-		return (NULL);
-	while (list->next)
-		list = list->next;
-	return (list);
-}
